requirements.txt
GitPython>=3.1.0
openai>=1.0.0
rich>=12.0.0
PyYAML>=6.0
pydantic>=1.10.0
termcolor>=2.3.0
